requirements.txt
aiogram~=2.14
environs~=8.0.0
pypdfium2~=3.20.0
pathlib~=1.0.1
Pillow~=9.4.0
asyncpg~=0.27.0
pdf2image
fpdf